+++ lib/run/browsers/base.js
@@ -2,7 +2,7 @@ import os from 'os';
 import path from 'path';
 import { when } from '@bigtest/convergence';
 
-import { mkdir, rimraf, writeFile } from '../util/common';
+import { mkdir, rimraf, writeFile } from '../../util/fs';
 import ChildProcess from '../process';
 
 export default class BaseBrowser extends ChildProcess {
