+++ lib/run/coordinator.js
@@ -1,6 +1,6 @@
 import { when } from '@bigtest/convergence';
 
-import logger from './logger';
+import logger from '../util/logger';
 import ClientServer from './client';
 import SocketServer from './sockets';
 import ProxyServer from './proxy';
