+++ lib/run/process.js
@@ -1,11 +1,8 @@
 import { spawn } from 'child_process';
 import { when } from '@bigtest/convergence';
 
-import {
-  pathExists,
-  binExists,
-  hasDescriptor
-} from './util/common';
+import { pathExists, binExists } from '../util/fs';
+import { hasDescriptor } from './util/common';
 
 const { assign } = Object;
