+++ lib/run/util/require.js
@@ -1,9 +1,9 @@
 import path from 'path';
 
+import { pathExists } from '../../util/fs';
 import Browser from '../browsers/base';
 import Reporter from '../reporters/base';
 import Plugin from '../plugins/base';
-import { pathExists } from './common';
 
 /**
  * Checks for the a file in a local directory within this package
