+++ lib/util/logger.js
